requirements.txt
Django>=2.2.3<2.3.0
djangorestframework==3.9.2
psycopg2